Rename types.ts to FormatOptions.ts
